codeblockpc[] was only partially initialised. Also change the type to array of uint32_t, and use memset() to initialise blocks[]
--HG--
branch : trunk

+++ src/codegen_x86.h
@@ -5,7 +5,7 @@
 
 //unsigned char codeblock[3][0x1000][1600];
 extern unsigned char rcodeblock[BLOCKS][1792+512+64];
-extern unsigned long codeblockpc[0x8000];
+extern uint32_t codeblockpc[0x8000];
 extern int codeblocknum[0x8000];
 
 extern uint8_t flaglookup[16][16];
